generate import shares agnostic to mrio, resolves #53; requires making region assignments based on imports data separate from MRIO country assignments
